fix: minor details
